fix bug to get min coordinates in the imaging session.
Since constructor of GoDBCoordinates initialized all coord values to 0,
and to take the min it was comparing image coordinate to min and max
(at the beginning initialized to 0), it was resulting to 0 coord min
in all cases.
Now min and max are directly provided my the multi index container
for each dimension.