WIP: Compute mesh attributes on the different channels.
There is still a bug: we need to get the channel name for the DB, in
order to save corresponding value. Channel name was taken from the
NavigationDockWidget, but this one has been changed and what was used
to provide the channel information becomes time point in the doppler
view case.